Add files via upload
The script tags are used to load app. js into the page so that it can be executed when someone clicks on the button labeled “generateQr()”. The code will generate a QR code with the text “Paste an URL or enter text, then press enter” and a button that when clicked on will generate the QR code.